+++ src/extensions/codemirror/CodeMirror.stories.tsx
@@ -1,6 +1,8 @@
 import React from "react";
 import { Meta, StoryFn } from "@storybook/react";
 
+import { helpersArgTypes } from "../../../.storybook/helpers";
+
 import { CodeEditor } from "./CodeMirror";
 
 export default {
@@ -10,6 +12,9 @@ export default {
     argTypes: {
         onChange: {
             action: "value changed",
+        },
+        intent: {
+            ...helpersArgTypes.exampleIntent,
         },
     },
 } as Meta<typeof CodeEditor>;
@@ -29,4 +34,21 @@ LinterExample.args = {
     defaultValue: "**test me**",
     mode: "javascript",
     useLinting: true,
+    autoFocus: true,
+};
+
+export const DisabledExample = TemplateFull.bind({});
+DisabledExample.args = {
+    name: "codeinput",
+    defaultValue: "**test me**",
+    mode: "javascript",
+    disabled: true,
+};
+
+export const IntentExample = TemplateFull.bind({});
+IntentExample.args = {
+    name: "codeinput",
+    defaultValue: "**test me**",
+    mode: "javascript",
+    intent: "warning",
 };